pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bm-gateway"
version = "0.1.0"
description = "Battery monitor gateway mono-repo with Python CLI/service, Home Assistant integration, Raspberry Pi setup, and web interface"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Marco Massari Calderone", email = "marco@marcomc.com" }
]
dependencies = [
  "paho-mqtt>=2.1.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.0",
  "pytest>=8.3.0",
  "ruff>=0.11.0",
]

[project.scripts]
bm-gateway = "bm_gateway.cli:main"

[tool.setuptools.package-dir]
"" = "python/src"

[tool.setuptools.packages.find]
where = ["python/src"]

[tool.pytest.ini_options]
testpaths = ["python/tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.11"
files = ["python/src", "python/tests"]
strict = true
